Updated with \env file

+++ src/comments/comments.controller.ts
@@ -7,10 +7,15 @@ import {
   Param,
   Delete,
   Query,
+  ForbiddenException,
+  NotFoundException,
+  UseGuards,
+  Req,
 } from '@nestjs/common';
 import { CommentsService } from './comments.service';
 import { CreateCommentDto } from './dto/create-comment.dto';
 import { UpdateCommentDto } from './dto/update-comment.dto';
+import { AuthGuard } from 'src/auth/auth.guard';
 
 @Controller('comments')
 export class CommentsController {
@@ -40,11 +45,27 @@ export class CommentsController {
   update(@Param('id') id: string, @Body() dto: UpdateCommentDto) {
     return this.commentsService.update(id, dto);
   }
-
   @Delete(':id')
-  remove(@Param('id') id: string) {
+  @UseGuards(AuthGuard)
+  async remove(@Param('id') id: string, @Req() req: any) {
+    const comment = await this.commentsService.findById(id);
+    console.log(comment);
+    if (!comment) {
+      throw new NotFoundException('Comment not found');
+      
+    }
+  
+    const isOwner = comment.user.toString() === req.user.userId;
+    const isAdmin = req.user.role === 'admin';
+  
+    if (!isOwner && !isAdmin) {
+      throw new ForbiddenException('You are not allowed to delete this comment');
+      
+    }
+  
     return this.commentsService.remove(id);
   }
+  
   @Post(':id/like')
   async likeComment(@Param('id') id: string) {
     return this.commentsService.likeComment(id);
